+++ Changes.md
@@ -2,6 +2,7 @@ Version Information
 ===================
 Version 3.6.0.2
   1. Only one column when editing.
+  2. Change 'barchart' setting to 'chart' to facilitate having no chart at all.
 
 Version 3.6.0.1
   1. Apply MDL-64819.